Added common handlers for CRT and MFC patch functions.
Every new version of Visual Studio introduces a new version of the CRT and MFC
DLLs. This means we need to add a new set of patch functions for each new
version of Visual Studio. However, the code in each patch function is nearly
identical to all the others. The old way these functions were structured, it
resulted in a lot of duplicate code.

This commit adds a set of common handlers, one for each "kind" of function.
A given kind of function has the same signature, regardless of which DLL the
function is exported from. The individual patch functions (for which there
MUST be one for each function redirected from each DLL) then call these common
handlers, which results in a reduction of the amount of duplicated code.