Fix for Issue 8228.
So the problem was here the changes I made earlier that combined the duties for Callstack::dump and Callstack::Resolve to one function called Callstack::resolve. 

So I'm fixing this by reverting the two functions into two seperate functions:
1. Callstack::dump to print out callstacks to the specified outputs
2. Callstack::resolve to store the formatted callstack for later retrieval. 
These two methods are almost identical, so any changes made to one needs to be made to the other. 

NOTE: I'm not sure why this was failing, but seperating out the code makes it work on my massive application. 

I also fixed the print function in utility.cpp which was not protected against NULL strings.

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/vld/vld/trunk;C65202