Add a test which loads a module dynamically during runtime. The app, dynamic.exe loads dynamic.dll during runtime and calls into an exported method. I also added some code in there to test for aggregating callstacks (using the INI file to control various options).
What I found was that any memory leaks that actually leaked in the module itself were not reported. That is a bug which I will be reporting on later.

All three tests combined push the code coverage up to 53% functions coverage, with 43% condition coverage. 
There is still a ways to go as far as increasing the coverage, but these test apps provide a good start for now.

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/vld/vld/trunk;C65866