The main problem in the bug is that VLD ignores allocations in DLL's that don't directly import or depend on vld_xxx.dll. Wierd I know, but that's how it works. Hence I'm reverting my previous change in vldint.h, and solving this by merely calling the exported VLDEnableModule and passing in the module handle of my dynamically loaded DLL's.
So the rule is:
If you dynamically load a module in your process space via LoadLibrary that does NOT link against vld_xxx.dll, then you must call VLDEnableModule and pass in the module handle for that DLL.

Changes:
--------------------------------
vld.cpp
Altered VisualLeakDetector::IsModuleExcluded so that it just loops though moduleinfo_t's and directly checks if the address (which is not actually the base address) is in the range of the module.
The assumption previously was that UINT_PTR address was the module base address, which it was not. It is merely the address of a stack frame. So here I check if it's higher than the base address, and lower than the high-address.
I also removed VisualLeakDetector::GetModuleName which was completely unused. And since it's not good to have dead, uncalled, untested code here I took it out.
If anyone needs it, they can look in the version control, and retrieve it.
--------------------------------
vldint.h
Reverted my last change to moduleinfo_s::operator<.
--------------------------------
testsuite.cpp
Put in a better way to wait for the threads to finish, instead of spinning wastefully in a loop. This uses WaitForMultipleObjects.
--------------------------------
Other test suites:
--------------------------------
I also updated a few of the other test suites.
With these changes, the dynamic.exe test now correctly leaks exactly 31 memory allocations.

Tested by running all test suites in 32 and 64 bit debug builds.

As for Issue 8512, I cannot reproduce it anymore with these changes.

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/vld/vld/trunk;C66054