Resolved the internal memory leaks in LoadTests.cpp. What the problem was that the test unloaded the MFC Dll by calling FreeLibrary. The problem is that VLD hooks load library, but doesn't hook unload library. I don't know exactly what is going on here, but until I debug it, I"m turning off the code to unload the mfc dll. Now running dynamic_app in 64 bit mode, will report the correct 31 memory leaks with no extra warnings about internal memory leaks in VLD.
vldint.h
I also removed the typedef structs, as the microsoft C++ compiler can recognize a simple struct definition without resorting to confusing typdefs. 
Also, the thread local storage types (tls_t) were kept in a set. However sets are supposed to be immutable, which this set is not treated that way. Therefore replaced the set with a map. 

TESTS:
Ran all the tests, and got the expected amount of memory leaks for each test.

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/vld/vld/trunk;C66166