Problem: Uncomment out the code in RunMFCLoaderTests. Rerun the test dynamic_app.exe
Result:
It will leak internal memory due to resolve strings not getting freed.

The problem was made worse in that m_Resolved was a WCHAR**, which required a lot of allocations. Now I replaced with a WCHAR* and I simply concatenate strings to the end of it.
Dump resolved call stacks looks just the same, only it's alot faster now.

This does not resolve the internal memory leak, it just reduces the amount of strings that do get leaked. Using these changes, it should be a bit easier to identify the culprit.

What I have found is that of all the callstacks that get resolved, most do correctly get free'd. Running dynamic_app.exe will leak 18 strings in both 32 and 64 bit configurations. (Remember, this is only when you FIRST resolve the callstacks, and then unload the MFC based dll).


git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/vld/vld/trunk;C66187