Adding Heap Validation to VLD.
One especially problematic way to crash an application is to allocate memory in one heap, and free it in another.
The MS debug heaps do detect this, but they do NOT report any callstacks indicating the allocation callstack. The MS debug heaps simply report an error condition, and issue a break leaving the end user guessing what went wrong. 

Visual leak detector can now pinpoint EXACTLY report on such a situation by reporting callstacks of the allocation and deallocation of the memory that corrupted the heap.

If a memory allocation is unmapped with the wrong heap handle, VLD will do a search through it's internal maps of memory allocations searching for the memory address. It will also find the heap handle that IS associated with the memory block.
This of course comes at a cost, as there are many heap allocs and heap frees that VLD does not monitor. To seperate the two, and keep performance similar to what it was before, this extra checking must be explicitely be turned on via INI file, or by setting:
VLD_OPT_VALIDATE_HEAPFREE
via VLDSetOptions.

I also wrote a unit test to verify this extra behavior. This test is called corruptions, and will be the test bed for further testing in VLD.

I also cleaned up the silly semantics of the strapp method (what a stupid name), renaming it StringAppend, and having it return a string, instead of passing in a string by reference. Which goes to prove that trying to be really clever just makes everyone confused.

TESTS:
Ran all tests in 32 bit, and all passed.

git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/vld/vld/trunk;C66462